Refactor current main page setup into a <NavigationWrapper /> component
