+++ frontend/src/Views/Home/HomePage.tsx
@@ -1,40 +1,17 @@
 import React from "react";
-import Header from "./Header";
+import NavigationWrapper from "../../Components/NavigationWrapper/NavigationWrapper";
 import SimpleLineChart from "./SimpleLineChart";
 
 import {
   createStyles,
-  Drawer,
-  Grid,
-  Paper,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TablePagination,
-  TableRow,
   Theme,
   Typography,
   withStyles,
   WithStyles
 } from "@material-ui/core";
 
-import {
-  Bar,
-  BarChart,
-  CartesianGrid,
-  Legend,
-  ResponsiveContainer,
-  Tooltip,
-  XAxis,
-  YAxis
-} from "recharts";
-
 const styles = (theme: Theme) =>
   createStyles({
-    root: {
-      display: "flex"
-    },
     appBarSpacer: theme.mixins.toolbar,
     chartContainer: {
       marginLeft: -22
@@ -44,9 +21,6 @@ const styles = (theme: Theme) =>
       height: "100vh",
       overflow: "auto",
       padding: theme.spacing.unit * 3
-    },
-    toolbar: {
-      paddingRight: 24
     }
   });
 
@@ -54,8 +28,7 @@ export interface IProps extends WithStyles<typeof styles> {}
 
 const HomePage = ({ classes }: IProps) => {
   return (
-    <div className={classes.root}>
-      <Header />
+    <NavigationWrapper>
       <main className={classes.content}>
         <div className={classes.appBarSpacer} />
         <h1>Monthly Sales Data</h1>
@@ -63,7 +36,7 @@ const HomePage = ({ classes }: IProps) => {
           <SimpleLineChart />
         </Typography>
       </main>
-    </div>
+    </NavigationWrapper>
   );
 };
